feat: upgrade cookie to 0.1.2
