+++ solagent-plugins/cookie/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "solagent-plugin-cookie"
-version = "0.1.1"
+version = "0.1.2"
 edition = "2021"
 authors = ["zTgx <beautifularea@gmail.com>"]
 repository = "https://github.com/zTgx/solagent.rs"
@@ -9,7 +9,7 @@ license = "Apache-2.0"
 description = "solagent plugin cookie"
 
 [dependencies]
-solagent-core = "0.1.4"
+solagent-core = "0.1.6"
 serde = { version = "1.0", features = ["derive"] }
 base64 = "0.22.1"
 reqwest = { version = "0.12", features = ["json"] }
