+++ solagent-plugins/cookie/rust-toolchain
@@ -1,2 +1,2 @@
 [toolchain]
-channel = "1.82.0"
+channel = "1.83.0"
